+++ test/samovar/options.rb
@@ -161,7 +161,7 @@ describe Samovar::Options do
 		end
 	end
 	
-	with "equals sign syntax" do
+	with "argument transform" do
 		let(:options) do
 			subject.parse do
 				option "-c/--config <path>", "The configuration file path."
@@ -170,50 +170,24 @@ describe Samovar::Options do
 			end
 		end
 		
-		it "parses a value given with an equals sign" do
-			values = options.parse(["--config=app.yml"], {})
+		it "rewrites selected equals syntax into argv pairs" do
+			values = options.parse(Samovar::Arguments.transform(["--config=app.yml"], keys: ["--config"]), {})
 			expect(values[:config]).to be == "app.yml"
 		end
 		
-		it "is equivalent to the space-separated form" do
-			expect(options.parse(["--config=app.yml"], {})).to be == options.parse(["--config", "app.yml"], {})
+		it "only transforms the selected keys" do
+			arguments = Samovar::Arguments.transform(["--config=app.yml", "--verbose=anything"], keys: ["--config"])
+			expect(arguments).to be == ["--config", "app.yml", "--verbose=anything"]
 		end
 		
 		it "splits on the first equals sign only" do
-			values = options.parse(["--config=a=b=c"], {})
-			expect(values[:config]).to be == "a=b=c"
+			arguments = Samovar::Arguments.transform(["--config=a=b=c"], keys: ["--config"])
+			expect(arguments).to be == ["--config", "a=b=c"]
 		end
 		
-		it "accepts an empty value after the equals sign" do
-			values = options.parse(["--config="], {})
-			expect(values[:config]).to be == ""
-		end
-		
-		it "parses an alternative prefix with an equals sign" do
-			values = options.parse(["-c=app.yml"], {})
-			expect(values[:config]).to be == "app.yml"
-		end
-		
-		it "treats a presence flag with an equals sign as set" do
-			values = options.parse(["--verbose=anything"], {})
-			expect(values[:verbose]).to be == true
-		end
-		
-		it "parses false-like boolean values case-insensitively" do
-			["false", "FALSE", "No", "off", "0"].each do |value|
-				expect(options.parse(["--debug=#{value}"], {})[:debug]).to be == false
-			end
-		end
-		
-		it "treats other boolean values as true" do
-			["true", "yes", "on", "1", "anything"].each do |value|
-				expect(options.parse(["--debug=#{value}"], {})[:debug]).to be == true
-			end
-		end
-		
-		it "inverts the value for a negated boolean flag" do
-			expect(options.parse(["--no-debug=off"], {})[:debug]).to be == true
-			expect(options.parse(["--no-debug=true"], {})[:debug]).to be == false
+		it "preserves empty values for explicit transforms" do
+			arguments = Samovar::Arguments.transform(["--config="], keys: ["--config"])
+			expect(arguments).to be == ["--config", ""]
 		end
 	end
 end